requirements.txt
APScheduler==3.11.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.2
cssselect2==0.8.0
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.2.11
django-apscheduler==0.7.0
django-auditlog==3.1.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-haystack==3.3.0
django-import-export==4.3.7
django-jsonfield==1.4.1
django-mathfilters==1.0.0
django-microsoft-auth==3.0.1
django-model-utils==5.0.0
django-simple-history==3.8.0
django-storages==1.14.6
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
future==1.0.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.0.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.2.6
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.2.3
pdfkit==1.0.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyHanko==0.27.1
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.2.3
pypdf==5.5.0
pypng==0.20220715.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzk==0.9
qrcode==8.2
reportlab==4.0.9
requests==2.32.3
requests-oauthlib==2.0.0
responses==0.25.7
rsa==4.9.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
svglib==1.5.1
swapper==1.4.0
tablib==3.8.0
tinycss2==1.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==5.0.0
urllib3==2.4.0
utils==1.0.2
webencodings==0.5.1
whitenoise==6.9.0
Whoosh==2.7.4
xhtml2pdf==0.2.17
XlsxWriter==3.2.3
